requirements.txt
torch
numpy
scipy
pandas
scikit-learn
mne
osl-ephys
dask[distributed]
matplotlib
seaborn
einops
tqdm
transformers
tokenizers
torchmetrics
torchview
vector_quantize_pytorch
wordfreq
beartype
pytorch-lightning
tensorboard
pyyaml
pytest
black
